requirements.txt
requests
pillow
google-generativeai